implemented product order relation and orders are now successfully  created in database

+++ MyStore.Services/Utils/Mapper.cs
@@ -66,7 +66,7 @@ namespace MyStore.Services.Utils
         {
             return new Order
             {
-                Price = dto.Price
+                Price = (decimal)dto.Price
             };
         }
     }
